Crate google-sqladmin1_beta4

This documentation was generated from SQL Admin crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

backup runs

get and list

databases

delete, get, insert, list, patch and update

flags

list

instances

clone, delete, export, get, import, insert, list, patch, promote replica, reset ssl config, restart, restore backup, start replica, stop replica and update

operations

get and list

ssl certs

delete, get, insert and list

tiers

list

users

delete, insert, list and update

Everything else about the SQL Admin v1_beta4 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.ssl_certs().delete(...).doit()
let r = hub.users().delete(...).doit()
let r = hub.databases().delete(...).doit()
let r = hub.databases().patch(...).doit()
let r = hub.instances().reset_ssl_config(...).doit()
let r = hub.instances().promote_replica(...).doit()
let r = hub.databases().update(...).doit()
let r = hub.operations().list(...).doit()
let r = hub.users().update(...).doit()
let r = hub.databases().insert(...).doit()
let r = hub.instances().patch(...).doit()
let r = hub.instances().clone(...).doit()
let r = hub.instances().delete(...).doit()
let r = hub.operations().get(...).doit()
let r = hub.instances().stop_replica(...).doit()
let r = hub.instances().start_replica(...).doit()
let r = hub.users().insert(...).doit()
let r = hub.instances().insert(...).doit()
let r = hub.instances().import(...).doit()
let r = hub.instances().update(...).doit()
let r = hub.instances().restart(...).doit()
let r = hub.instances().export(...).doit()
let r = hub.instances().restore_backup(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-sqladmin1_beta4 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-sqladmin1_beta4" as sqladmin1_beta4;
use sqladmin1_beta4::User;
use sqladmin1_beta4::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: User = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().update(&req, "project", "instance", "host", "name")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AclEntry
An entry for an Access Control list.

BackupConfiguration
Database instance backup configuration.

BackupRun
A database instance backup run resource.

BackupRunGetCallBuilder
Retrieves a resource containing information about a backup run.

BackupRunListCallBuilder
Lists all backup runs associated with a given instance and configuration in the reverse chronological order of the enqueued time.

BackupRunMethodsBuilder
A builder providing access to all methods supported on backupRun resources. It is not used directly, but through the SQLAdmin hub.

BackupRunsListResponse
Backup run list results.

BinLogCoordinates
Binary log coordinates.

CloneContext
Database instance clone context.

Database
A database resource inside a Cloud SQL instance.

DatabaseDeleteCallBuilder
Deletes a resource containing information about a database inside a Cloud SQL instance.

DatabaseFlags
MySQL flags for Cloud SQL instances.

DatabaseGetCallBuilder
Retrieves a resource containing information about a database inside a Cloud SQL instance.

DatabaseInsertCallBuilder
Inserts a resource containing information about a database inside a Cloud SQL instance.

DatabaseInstance
A Cloud SQL instance resource.

DatabaseListCallBuilder
Lists databases in the specified Cloud SQL instance.

DatabaseMethodsBuilder
A builder providing access to all methods supported on database resources. It is not used directly, but through the SQLAdmin hub.

DatabasePatchCallBuilder
Updates a resource containing information about a database inside a Cloud SQL instance. This method supports patch semantics.

DatabaseUpdateCallBuilder
Updates a resource containing information about a database inside a Cloud SQL instance.

DatabasesListResponse
Database list response.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ExportContext
Database instance export context.

ExportContextCsvExportOptions
Options for exporting data as CSV.

ExportContextSqlExportOptions
Options for exporting data as SQL statements.

Flag
A Google Cloud SQL service flag resource.

FlagListCallBuilder
List all available database flags for Google Cloud SQL instances.

FlagMethodsBuilder
A builder providing access to all methods supported on flag resources. It is not used directly, but through the SQLAdmin hub.

FlagsListResponse
Flags list response.

ImportContext
Database instance import context.

ImportContextCsvImportOptions
Options for importing data as CSV.

InstanceCloneCallBuilder
Creates a Cloud SQL instance as a clone of the source instance.

InstanceDeleteCallBuilder
Deletes a Cloud SQL instance.

InstanceExportCallBuilder
Exports data from a Cloud SQL instance to a Google Cloud Storage bucket as a MySQL dump file.

InstanceGetCallBuilder
Retrieves a resource containing information about a Cloud SQL instance.

InstanceImportCallBuilder
Imports data into a Cloud SQL instance from a MySQL dump file in Google Cloud Storage.

InstanceInsertCallBuilder
Creates a new Cloud SQL instance.

InstanceListCallBuilder
Lists instances under a given project in the alphabetical order of the instance name.

InstanceMethodsBuilder
A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

InstancePatchCallBuilder
Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.. This method supports patch semantics.

InstancePromoteReplicaCallBuilder
Promotes the read replica instance to be a stand-alone Cloud SQL instance.

InstanceResetSslConfigCallBuilder
Deletes all client certificates and generates a new server SSL certificate for the instance. The changes will not take effect until the instance is restarted. Existing instances without a server certificate will need to call this once to set a server certificate.

InstanceRestartCallBuilder
Restarts a Cloud SQL instance.

InstanceRestoreBackupCallBuilder
Restores a backup of a Cloud SQL instance.

InstanceStartReplicaCallBuilder
Starts the replication in the read replica instance.

InstanceStopReplicaCallBuilder
Stops the replication in the read replica instance.

InstanceUpdateCallBuilder
Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.

InstancesCloneRequest
Database instance clone request.

InstancesExportRequest
Database instance export request.

InstancesImportRequest
Database instance import request.

InstancesListResponse
Database instances list response.

InstancesRestoreBackupRequest
Database instance restore backup request.

IpConfiguration
IP Management configuration.

IpMapping
Database instance IP Mapping.

LocationPreference
Preferred location. This specifies where a Cloud SQL instance should preferably be located, either in a specific Compute Engine zone, or co-located with an App Engine application. Note that if the preferred location is not available, the instance will be located as close as possible within the region. Only one location may be specified.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Operation
An Operations resource contains information about database instance operations such as create, delete, and restart. Operations resources are created in response to operations that were initiated; you never create them directly.

OperationError
Database instance operation error.

OperationErrorNested
If errors occurred during processing of this operation, this field will be populated.

OperationGetCallBuilder
Retrieves an instance operation that has been performed on an instance.

OperationListCallBuilder
Lists all instance operations that have been performed on the given Cloud SQL instance in the reverse chronological order of the start time.

OperationMethodsBuilder
A builder providing access to all methods supported on operation resources. It is not used directly, but through the SQLAdmin hub.

OperationsListResponse
Database instance list operations response.

RestoreBackupContext
Database instance restore from backup context.

SQLAdmin
Central instance to access all SQLAdmin related resource activities

Settings
Database instance settings.

SslCert
SslCerts Resource

SslCertDeleteCallBuilder
Deletes the SSL certificate. The change will not take effect until the instance is restarted.

SslCertDetail
SslCertDetail.

SslCertGetCallBuilder
Retrieves a particular SSL certificate. Does not include the private key (required for usage). The private key must be saved from the response to initial creation.

SslCertInsertCallBuilder
Creates an SSL certificate and returns it along with the private key and server certificate authority. The new certificate will not be usable until the instance is restarted.

SslCertListCallBuilder
Lists all of the current SSL certificates for the instance.

SslCertMethodsBuilder
A builder providing access to all methods supported on sslCert resources. It is not used directly, but through the SQLAdmin hub.

SslCertsInsertRequest
SslCerts insert request.

SslCertsInsertResponse
SslCert insert response.

SslCertsListResponse
SslCerts list response.

Tier
A Google Cloud SQL service tier resource.

TierListCallBuilder
Lists all available service tiers for Google Cloud SQL, for example D1, D2. For related information, see Pricing.

TierMethodsBuilder
A builder providing access to all methods supported on tier resources. It is not used directly, but through the SQLAdmin hub.

TiersListResponse
Tiers list response.

User
A Cloud SQL user resource.

UserDeleteCallBuilder
Deletes a user from a Cloud SQL instance.

UserInsertCallBuilder
Creates a new user in a Cloud SQL instance.

UserListCallBuilder
Lists users in the specified Cloud SQL instance.

UserMethodsBuilder
A builder providing access to all methods supported on user resources. It is not used directly, but through the SQLAdmin hub.

UserUpdateCallBuilder
Updates an existing user in a Cloud SQL instance.

UsersListResponse
User list response.

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.