Module google_sqladmin1_beta4::api

Structs

AclEntry
An entry for an Access Control list.

ApiWarning
An Admin API warning message.

BackupConfiguration
Database instance backup configuration.

BackupContext
Backup context.

BackupRetentionSettings
We currently only support backup retention by specifying the number of backups we will retain.

BackupRun
A BackupRun resource.

BackupRunDeleteCall
Deletes the backup taken by a backup run.

BackupRunGetCall
Retrieves a resource containing information about a backup run.

BackupRunInsertCall
Creates a new backup run on demand. This method is applicable only to Second Generation instances.

BackupRunListCall
Lists all backup runs associated with a given instance and configuration in the reverse chronological order of the backup initiation time.

BackupRunMethods
A builder providing access to all methods supported on backupRun resources. It is not used directly, but through the SQLAdmin hub.

BackupRunsListResponse
Backup run list results.

BinLogCoordinates
Binary log coordinates.

CloneContext
Database instance clone context.

Database
Represents a SQL database on the Cloud SQL instance.

DatabaseDeleteCall
Deletes a database from a Cloud SQL instance.

DatabaseFlags
Database flags for Cloud SQL instances.

DatabaseGetCall
Retrieves a resource containing information about a database inside a Cloud SQL instance.

DatabaseInsertCall
Inserts a resource containing information about a database inside a Cloud SQL instance.

DatabaseInstance
A Cloud SQL instance resource.

DatabaseInstanceFailoverReplica
The name and status of the failover replica. This property is applicable only to Second Generation instances.

DatabaseListCall
Lists databases in the specified Cloud SQL instance.

DatabaseMethods
A builder providing access to all methods supported on database resources. It is not used directly, but through the SQLAdmin hub.

DatabasePatchCall
Partially updates a resource containing information about a database inside a Cloud SQL instance. This method supports patch semantics.

DatabaseUpdateCall
Updates a resource containing information about a database inside a Cloud SQL instance.

DatabasesListResponse
Database list response.

DemoteMasterConfiguration
Read-replica configuration for connecting to the on-premises primary instance.

DemoteMasterContext
Database instance demote primary instance context.

DemoteMasterMySqlReplicaConfiguration
Read-replica configuration specific to MySQL databases.

DenyMaintenancePeriod
Deny Maintenance Periods. This specifies a date range during when all CSA rollout will be denied.

DiskEncryptionConfiguration
Disk encryption configuration for an instance.

DiskEncryptionStatus
Disk encryption status for an instance.

ExportContext
Database instance export context.

ExportContextCsvExportOptions
Options for exporting data as CSV. MySQL and PostgreSQL instances only.

ExportContextSqlExportOptions
Options for exporting data as SQL statements.

ExportContextSqlExportOptionsMysqlExportOptions
Options for exporting from MySQL.

FailoverContext
Database instance failover context.

Flag
A flag resource.

FlagListCall
List all available database flags for Cloud SQL instances.

FlagMethods
A builder providing access to all methods supported on flag resources. It is not used directly, but through the SQLAdmin hub.

FlagsListResponse
Flags list response.

ImportContext
Database instance import context.

ImportContextBakImportOptions
Import parameters specific to SQL Server .BAK files

ImportContextBakImportOptionsEncryptionOptions
There is no detailed description.

ImportContextCsvImportOptions
Options for importing data as CSV.

InsightsConfig
Insights configuration. This specifies when Cloud SQL Insights feature is enabled and optional configuration.

InstanceAddServerCaCall
Add a new trusted Certificate Authority (CA) version for the specified instance. Required to prepare for a certificate rotation. If a CA version was previously added but never used in a certificate rotation, this operation replaces that version. There cannot be more than one CA version waiting to be rotated in.

InstanceCloneCall
Creates a Cloud SQL instance as a clone of the source instance. Using this operation might cause your instance to restart.

InstanceDeleteCall
Deletes a Cloud SQL instance.

InstanceDemoteMasterCall
Demotes the stand-alone instance to be a Cloud SQL read replica for an external database server.

InstanceExportCall
Exports data from a Cloud SQL instance to a Cloud Storage bucket as a SQL dump or CSV file.

InstanceFailoverCall
Failover the instance to its failover replica instance. Using this operation might cause your instance to restart.

InstanceGetCall
Retrieves a resource containing information about a Cloud SQL instance.

InstanceImportCall
Imports data into a Cloud SQL instance from a SQL dump or CSV file in Cloud Storage.

InstanceInsertCall
Creates a new Cloud SQL instance.

InstanceListCall
Lists instances under a given project.

InstanceListServerCaCall
Lists all of the trusted Certificate Authorities (CAs) for the specified instance. There can be up to three CAs listed: the CA that was used to sign the certificate that is currently in use, a CA that has been added but not yet used to sign a certificate, and a CA used to sign a certificate that has previously rotated out.

InstanceMethods
A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

InstancePatchCall
Updates settings of a Cloud SQL instance. This method supports patch semantics.

InstancePromoteReplicaCall
Promotes the read replica instance to be a stand-alone Cloud SQL instance. Using this operation might cause your instance to restart.

InstanceResetSslConfigCall
Deletes all client certificates and generates a new server SSL certificate for the instance.

InstanceRestartCall
Restarts a Cloud SQL instance.

InstanceRestoreBackupCall
Restores a backup of a Cloud SQL instance. Using this operation might cause your instance to restart.

InstanceRotateServerCaCall
Rotates the server certificate to one signed by the Certificate Authority (CA) version previously added with the addServerCA method.

InstanceStartReplicaCall
Starts the replication in the read replica instance.

InstanceStopReplicaCall
Stops the replication in the read replica instance.

InstanceTruncateLogCall
Truncate MySQL general and slow query log tables MySQL only.

InstanceUpdateCall
Updates settings of a Cloud SQL instance. Using this operation might cause your instance to restart.

InstancesCloneRequest
Database instance clone request.

InstancesDemoteMasterRequest
Database demote primary instance request.

InstancesExportRequest
Database instance export request.

InstancesFailoverRequest
Instance failover request.

InstancesImportRequest
Database instance import request.

InstancesListResponse
Database instances list response.

InstancesListServerCasResponse
Instances ListServerCas response.

InstancesRestoreBackupRequest
Database instance restore backup request.

InstancesRotateServerCaRequest
Rotate Server CA request.

InstancesTruncateLogRequest
Instance truncate log request.

IpConfiguration
IP Management configuration.

IpMapping
Database instance IP Mapping.

LocationPreference
Preferred location. This specifies where a Cloud SQL instance is located, either in a specific Compute Engine zone, or co-located with an App Engine application. Note that if the preferred location is not available, the instance will be located as close as possible within the region. Only one location may be specified.

MaintenanceWindow
Maintenance window. This specifies when a Cloud SQL instance is restarted for system maintenance purposes.

MySqlReplicaConfiguration
Read-replica configuration specific to MySQL databases.

OnPremisesConfiguration
On-premises instance configuration.

Operation
An Operation resource. For successful operations that return an Operation resource, only the fields relevant to the operation are populated in the resource.

OperationError
Database instance operation error.

OperationErrors
Database instance operation errors list wrapper.

OperationGetCall
Retrieves an instance operation that has been performed on an instance.

OperationListCall
Lists all instance operations that have been performed on the given Cloud SQL instance in the reverse chronological order of the start time.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the SQLAdmin hub.

OperationsListResponse
Database instance list operations response.

ProjectInstanceRescheduleMaintenanceCall
Reschedules the maintenance on the given instance.

ProjectInstanceStartExternalSyncCall
Start External primary instance migration.

ProjectInstanceVerifyExternalSyncSettingCall
Verify External primary instance external sync settings.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the SQLAdmin hub.

ReplicaConfiguration
Read-replica configuration for connecting to the primary instance.

Reschedule
There is no detailed description.

RestoreBackupContext
Database instance restore from backup context. Backup context contains source instance id and project id.

RotateServerCaContext
Instance rotate server CA context.

SQLAdmin
Central instance to access all SQLAdmin related resource activities

Settings
Database instance settings.

SqlActiveDirectoryConfig
Active Directory configuration, relevant only for Cloud SQL for SQL Server.

SqlExternalSyncSettingError
External primary instance migration setting error/warning.

SqlInstancesRescheduleMaintenanceRequestBody
Reschedule options for maintenance windows.

SqlInstancesVerifyExternalSyncSettingsResponse
Instance verify external sync settings response.

SqlScheduledMaintenance
Any scheduled maintenancce for this instance.

SqlServerDatabaseDetails
Represents a Sql Server database on the Cloud SQL instance.

SqlServerUserDetails
Represents a Sql Server user on the Cloud SQL instance.

SslCert
SslCerts Resource

SslCertCreateEphemeralCall
Generates a short-lived X509 certificate containing the provided public key and signed by a private key specific to the target instance. Users may use the certificate to authenticate as themselves when connecting to the database.

SslCertDeleteCall
Deletes the SSL certificate. For First Generation instances, the certificate remains valid until the instance is restarted.

SslCertDetail
SslCertDetail.

SslCertGetCall
Retrieves a particular SSL certificate. Does not include the private key (required for usage). The private key must be saved from the response to initial creation.

SslCertInsertCall
Creates an SSL certificate and returns it along with the private key and server certificate authority. The new certificate will not be usable until the instance is restarted.

SslCertListCall
Lists all of the current SSL certificates for the instance.

SslCertMethods
A builder providing access to all methods supported on sslCert resources. It is not used directly, but through the SQLAdmin hub.

SslCertsCreateEphemeralRequest
SslCerts create ephemeral certificate request.

SslCertsInsertRequest
SslCerts insert request.

SslCertsInsertResponse
SslCert insert response.

SslCertsListResponse
SslCerts list response.

Tier
A Google Cloud SQL service tier resource.

TierListCall
Lists all available machine types (tiers) for Cloud SQL, for example, db-custom-1-3840. For related information, see Pricing.

TierMethods
A builder providing access to all methods supported on tier resources. It is not used directly, but through the SQLAdmin hub.

TiersListResponse
Tiers list response.

TruncateLogContext
Database Instance truncate log context.

User
A Cloud SQL user resource.

UserDeleteCall
Deletes a user from a Cloud SQL instance.

UserInsertCall
Creates a new user in a Cloud SQL instance.

UserListCall
Lists users in the specified Cloud SQL instance.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the SQLAdmin hub.

UserUpdateCall
Updates an existing user in a Cloud SQL instance.

UsersListResponse
User list response.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.