Enum google-sqladmin1_beta4::cmn::Result

pub enum Result<T = ()> {
    HttpError(HttpError),
    MissingAPIKey,
    MissingToken,
    FieldClash(&'static str),
    Failure(Response),
    Success(T),
}

A universal result type used as return for all action method results.

Variants

HttpError

The http connection failed

MissingAPIKey

We needed an API key for authentication, but didn't obtain one. Neither through the authenticator, nor through the Delegate.

MissingToken

We required a Token, but didn't get one from the Authenticator

FieldClash

An additional, free form field clashed with one of the built-in optional ones

Failure

Indicates an HTTP repsonse with a non-success status code

Success

It worked !