Struct google_sqladmin1_beta4::BackupConfiguration

pub struct BackupConfiguration {
    pub kind: Option<String>,
    pub enabled: Option<bool>,
    pub binary_log_enabled: Option<bool>,
    pub start_time: Option<String>,
}

Database instance backup configuration.

This type is not used in any activity, and only used as part of another schema.

Fields
kind: Option<String>

This is always sql#backupConfiguration.

enabled: Option<bool>

Whether this configuration is enabled.

binary_log_enabled: Option<bool>

Whether binary log is enabled. If backup configuration is disabled, binary log must be disabled as well.

start_time: Option<String>

Start time for the daily backup configuration in UTC timezone in the 24 hour format - HH:MM.