Struct google-sqladmin1_beta4::InstanceMethods

pub struct InstanceMethods<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-sqladmin1_beta4" as sqladmin1_beta4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clone(...)`, `delete(...)`, `export(...)`, `get(...)`, `import(...)`, `insert(...)`, `list(...)`, `patch(...)`, `promote_replica(...)`, `reset_ssl_config(...)`, `restart(...)`, `restore_backup(...)`, `start_replica(...)`, `stop_replica(...)` and `update(...)`
// to build up your call.
let rb = hub.instances();

Methods

impl<'a, C, NC, A> InstanceMethods<'a, C, NC, A>

fn reset_ssl_config(&self, project: &str, instance: &str) -> InstanceResetSslConfigCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes all client certificates and generates a new server SSL certificate for the instance. The changes will not take effect until the instance is restarted. Existing instances without a server certificate will need to call this once to set a server certificate.

fn promote_replica(&self, project: &str, instance: &str) -> InstancePromoteReplicaCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Promotes the read replica instance to be a stand-alone Cloud SQL instance.

fn get(&self, project: &str, instance: &str) -> InstanceGetCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a resource containing information about a Cloud SQL instance.

fn patch(&self, request: &DatabaseInstance, project: &str, instance: &str) -> InstancePatchCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.. This method supports patch semantics.

fn restart(&self, project: &str, instance: &str) -> InstanceRestartCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Restarts a Cloud SQL instance.

fn delete(&self, project: &str, instance: &str) -> InstanceDeleteCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a Cloud SQL instance.

fn stop_replica(&self, project: &str, instance: &str) -> InstanceStopReplicaCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Stops the replication in the read replica instance.

fn start_replica(&self, project: &str, instance: &str) -> InstanceStartReplicaCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Starts the replication in the read replica instance.

fn insert(&self, request: &DatabaseInstance, project: &str) -> InstanceInsertCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a new Cloud SQL instance.

fn list(&self, project: &str) -> InstanceListCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists instances under a given project in the alphabetical order of the instance name.

fn import(&self, request: &InstancesImportRequest, project: &str, instance: &str) -> InstanceImportCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Imports data into a Cloud SQL instance from a MySQL dump file in Google Cloud Storage.

fn update(&self, request: &DatabaseInstance, project: &str, instance: &str) -> InstanceUpdateCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.

fn clone(&self, request: &InstancesCloneRequest, project: &str, instance: &str) -> InstanceCloneCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a Cloud SQL instance as a clone of the source instance.

fn export(&self, request: &InstancesExportRequest, project: &str, instance: &str) -> InstanceExportCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Exports data from a Cloud SQL instance to a Google Cloud Storage bucket as a MySQL dump file.

fn restore_backup(&self, request: &InstancesRestoreBackupRequest, project: &str, instance: &str) -> InstanceRestoreBackupCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Restores a backup of a Cloud SQL instance.