Struct google-sqladmin1_beta4::IpConfiguration

pub struct IpConfiguration {
    pub require_ssl: Option<bool>,
    pub ipv4_enabled: Option<bool>,
    pub authorized_networks: Vec<AclEntry>,
}

IP Management configuration.

This type is not used in any activity, and only used as part of another schema.

Fields

require_ssl

Whether the mysqld should default to 'REQUIRE X509' for users connecting over IP.

ipv4_enabled

Whether the instance should be assigned an IP address or not.

authorized_networks

The list of external networks that are allowed to connect to the instance using the IP. In CIDR notation, also known as 'slash' notation (e.g. 192.168.100.0/24).