Struct google-sqladmin1_beta4::Operation

pub struct Operation {
    pub status: Option<String>,
    pub import_context: Option<ImportContext>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub export_context: Option<ExportContext>,
    pub target_project: Option<String>,
    pub target_id: Option<String>,
    pub operation_type: Option<String>,
    pub insert_time: Option<String>,
    pub target_link: Option<String>,
    pub start_time: Option<String>,
    pub error: Option<OperationErrorNested>,
    pub end_time: Option<String>,
    pub self_link: Option<String>,
    pub user: Option<String>,
}

An Operations resource contains information about database instance operations such as create, delete, and restart. Operations resources are created in response to operations that were initiated; you never create them directly.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete ssl certs (response)

delete users (response)

delete databases (response)

patch databases (response)

reset ssl config instances (response)

promote replica instances (response)

update databases (response)

list operations (none)

update users (response)

insert databases (response)

patch instances (response)

clone instances (response)

delete instances (response)

get operations (response)

stop replica instances (response)

start replica instances (response)

insert users (response)

insert instances (response)

import instances (response)

update instances (response)

restart instances (response)

export instances (response)

restore backup instances (response)

Fields

status

The status of an operation. Valid values are PENDING, RUNNING, DONE, UNKNOWN.

import_context

The context for import operation, if applicable.

kind

This is always sql#instanceOperation.

name

An identifier that uniquely identifies the operation. You can use this identifier to retrieve the Operations resource that has information about the operation.

export_context

The context for export operation, if applicable.

target_project

The project ID of the target instance related to this operation.

target_id

Name of the database instance related to this operation.

operation_type

TODO(b/18431310): update this list to reflect current values. The type of the operation. Valid values are CREATE, DELETE, UPDATE, RESTART, IMPORT, EXPORT, BACKUP_VOLUME, RESTORE_VOLUME.

insert_time

The time this operation was enqueued in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00.094Z.

target_link

The URI of the instance related to the operation.

start_time

The time this operation actually started in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00.094Z.

error

If errors occurred during processing of this operation, this field will be populated.

end_time

The time this operation finished in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00.094Z.

self_link

The URI of this resource.

user

The email address of the user who initiated this operation.