Struct google_sqladmin1_beta4::Settings

pub struct Settings {
    pub database_flags: Option<Vec<DatabaseFlags>>,
    pub kind: Option<String>,
    pub authorized_gae_applications: Option<Vec<String>>,
    pub activation_policy: Option<String>,
    pub backup_configuration: Option<BackupConfiguration>,
    pub ip_configuration: Option<IpConfiguration>,
    pub crash_safe_replication_enabled: Option<bool>,
    pub database_replication_enabled: Option<bool>,
    pub replication_type: Option<String>,
    pub tier: Option<String>,
    pub pricing_plan: Option<String>,
    pub settings_version: Option<String>,
    pub location_preference: Option<LocationPreference>,
}

Database instance settings.

This type is not used in any activity, and only used as part of another schema.

Fields

database_flags

The database flags passed to the instance at startup.

kind

This is always sql#settings.

authorized_gae_applications

The App Engine app IDs that can access this instance.

activation_policy

The activation policy for this instance. This specifies when the instance should be activated and is applicable only when the instance state is RUNNABLE. This can be one of the following. ALWAYS: The instance should always be active. NEVER: The instance should never be activated. ON_DEMAND: The instance is activated upon receiving requests.

backup_configuration

The daily backup configuration for the instance.

ip_configuration

The settings for IP Management. This allows to enable or disable the instance IP and manage which external networks can connect to the instance.

crash_safe_replication_enabled

Configuration specific to read replica instances. Indicates whether database flags for crash-safe replication are enabled.

database_replication_enabled

Configuration specific to read replica instances. Indicates whether replication is enabled or not.

replication_type

The type of replication this instance uses. This can be either ASYNCHRONOUS or SYNCHRONOUS.

tier

The tier of service for this instance, for example D1, D2. For more information, see pricing.

pricing_plan

The pricing plan for this instance. This can be either PER_USE or PACKAGE.

settings_version

The version of instance settings. This is a required field for update method to make sure concurrent updates are handled properly. During update, use the most recent settingsVersion value for this instance and do not try to update this value.

location_preference

The location preference settings. This allows the instance to be located as near as possible to either an App Engine app or GCE zone for better performance.