Struct google_sqladmin1_beta4::SslCertMethods

pub struct SslCertMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on sslCert resources. It is not used directly, but through the SQLAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_sqladmin1_beta4 as sqladmin1_beta4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create_ephemeral(...)`, `delete(...)`, `get(...)`, `insert(...)` and `list(...)`
// to build up your call.
let rb = hub.ssl_certs();

Methods

impl<'a, C, A> SslCertMethods<'a, C, A>

fn insert(&self, request: SslCertsInsertRequest, project: &str, instance: &str) -> SslCertInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an SSL certificate and returns it along with the private key and server certificate authority. The new certificate will not be usable until the instance is restarted.

Arguments

request - No description provided.

project - Project ID of the project to which the newly created Cloud SQL instances should belong.

instance - Cloud SQL instance ID. This does not include the project ID.

fn delete(&self, project: &str, instance: &str, sha1_fingerprint: &str) -> SslCertDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the SSL certificate. The change will not take effect until the instance is restarted.

Arguments

project - Project ID of the project that contains the instance to be deleted.

instance - Cloud SQL instance ID. This does not include the project ID.

sha1Fingerprint - Sha1 FingerPrint.

fn get(&self, project: &str, instance: &str, sha1_fingerprint: &str) -> SslCertGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a particular SSL certificate. Does not include the private key (required for usage). The private key must be saved from the response to initial creation.

Arguments

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

sha1Fingerprint - Sha1 FingerPrint.

fn create_ephemeral(&self, request: SslCertsCreateEphemeralRequest, project: &str, instance: &str) -> SslCertCreateEphemeralCall<'a, C, A>

Create a builder to help you perform the following task:

Generates a short-lived X509 certificate containing the provided public key and signed by a private key specific to the target instance. Users may use the certificate to authenticate as themselves when connecting to the database.

Arguments

request - No description provided.

project - Project ID of the Cloud SQL project.

instance - Cloud SQL instance ID. This does not include the project ID.

fn list(&self, project: &str, instance: &str) -> SslCertListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all of the current SSL certificates for the instance.

Arguments

project - Project ID of the project for which to list Cloud SQL instances.

instance - Cloud SQL instance ID. This does not include the project ID.