Crate google-sqladmin1_beta4

This documentation was generated from SQL Admin crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

backup runs (get and list)

databases (delete, get, insert, list, patch and update)

flags (list)

instances (clone, delete, export, get, import, insert, list, patch, promote replica, reset ssl config, restart, restore backup, start replica, stop replica and update)

operations (get and list)

ssl certs (delete, get, insert and list)

tiers (list)

users (delete, insert, list and update)

Everything else about the SQL Admin v1_beta4 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.ssl_certs().delete(...).doit()
let r = hub.users().delete(...).doit()
let r = hub.databases().delete(...).doit()
let r = hub.databases().patch(...).doit()
let r = hub.instances().reset_ssl_config(...).doit()
let r = hub.instances().promote_replica(...).doit()
let r = hub.databases().update(...).doit()
let r = hub.operations().list(...).doit()
let r = hub.users().update(...).doit()
let r = hub.databases().insert(...).doit()
let r = hub.instances().patch(...).doit()
let r = hub.instances().clone(...).doit()
let r = hub.instances().delete(...).doit()
let r = hub.operations().get(...).doit()
let r = hub.instances().stop_replica(...).doit()
let r = hub.instances().start_replica(...).doit()
let r = hub.users().insert(...).doit()
let r = hub.instances().insert(...).doit()
let r = hub.instances().import(...).doit()
let r = hub.instances().update(...).doit()
let r = hub.instances().restart(...).doit()
let r = hub.instances().export(...).doit()
let r = hub.instances().restore_backup(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-sqladmin1_beta4 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-sqladmin1_beta4" as sqladmin1_beta4;
use sqladmin1_beta4::User;
use sqladmin1_beta4::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: User = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().update(&req, "project", "instance", "host", "name")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

AclEntry
An entry for an Access Control list.

This type is not used in any activity, and only used as part of another schema.

BackupConfiguration
Database instance backup configuration.

This type is not used in any activity, and only used as part of another schema.

BackupRun
A database instance backup run resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get backup runs (response)

list backup runs (none)

BackupRunGetMethodBuilder
Retrieves a resource containing information about a backup run.

BackupRunListMethodBuilder
Lists all backup runs associated with a given instance and configuration in the reverse chronological order of the enqueued time.

BackupRunMethodsBuilder
A builder providing access to all methods supported on backupRun resources. It is not used directly, but through the SQLAdmin hub.

BackupRunsListResponse
Backup run list results.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list backup runs (response)

BinLogCoordinates
Binary log coordinates.

This type is not used in any activity, and only used as part of another schema.

CloneContext
Database instance clone context.

This type is not used in any activity, and only used as part of another schema.

Database
A database resource inside a Cloud SQL instance.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get databases (response)

list databases (none)

patch databases (request)

insert databases (request)

delete databases (none)

update databases (request)

DatabaseDeleteMethodBuilder
Deletes a resource containing information about a database inside a Cloud SQL instance.

DatabaseFlags
MySQL flags for Cloud SQL instances.

This type is not used in any activity, and only used as part of another schema.

DatabaseGetMethodBuilder
Retrieves a resource containing information about a database inside a Cloud SQL instance.

DatabaseInsertMethodBuilder
Inserts a resource containing information about a database inside a Cloud SQL instance.

DatabaseInstance
A Cloud SQL instance resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch instances (request)

insert instances (request)

get instances (response)

update instances (request)

DatabaseListMethodBuilder
Lists databases in the specified Cloud SQL instance.

DatabaseMethodsBuilder
A builder providing access to all methods supported on database resources. It is not used directly, but through the SQLAdmin hub.

DatabasePatchMethodBuilder
Updates a resource containing information about a database inside a Cloud SQL instance. This method supports patch semantics.

DatabaseUpdateMethodBuilder
Updates a resource containing information about a database inside a Cloud SQL instance.

DatabasesListResponse
Database list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list databases (response)

ExportContext
Database instance export context.

This type is not used in any activity, and only used as part of another schema.

ExportContextCsvExportOptions
Options for exporting data as CSV.

This type is not used in any activity, and only used as part of another schema.

ExportContextSqlExportOptions
Options for exporting data as SQL statements.

This type is not used in any activity, and only used as part of another schema.

Flag
A Google Cloud SQL service flag resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list flags (none)

FlagListMethodBuilder
List all available database flags for Google Cloud SQL instances.

FlagMethodsBuilder
A builder providing access to all methods supported on flag resources. It is not used directly, but through the SQLAdmin hub.

FlagsListResponse
Flags list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list flags (response)

ImportContext
Database instance import context.

This type is not used in any activity, and only used as part of another schema.

ImportContextCsvImportOptions
Options for importing data as CSV.

This type is not used in any activity, and only used as part of another schema.

InstanceCloneMethodBuilder
Creates a Cloud SQL instance as a clone of the source instance.

InstanceDeleteMethodBuilder
Deletes a Cloud SQL instance.

InstanceExportMethodBuilder
Exports data from a Cloud SQL instance to a Google Cloud Storage bucket as a MySQL dump file.

InstanceGetMethodBuilder
Retrieves a resource containing information about a Cloud SQL instance.

InstanceImportMethodBuilder
Imports data into a Cloud SQL instance from a MySQL dump file in Google Cloud Storage.

InstanceInsertMethodBuilder
Creates a new Cloud SQL instance.

InstanceListMethodBuilder
Lists instances under a given project in the alphabetical order of the instance name.

InstanceMethodsBuilder
A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

InstancePatchMethodBuilder
Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.. This method supports patch semantics.

InstancePromoteReplicaMethodBuilder
Promotes the read replica instance to be a stand-alone Cloud SQL instance.

InstanceResetSslConfigMethodBuilder
Deletes all client certificates and generates a new server SSL certificate for the instance. The changes will not take effect until the instance is restarted. Existing instances without a server certificate will need to call this once to set a server certificate.

InstanceRestartMethodBuilder
Restarts a Cloud SQL instance.

InstanceRestoreBackupMethodBuilder
Restores a backup of a Cloud SQL instance.

InstanceStartReplicaMethodBuilder
Starts the replication in the read replica instance.

InstanceStopReplicaMethodBuilder
Stops the replication in the read replica instance.

InstanceUpdateMethodBuilder
Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.

InstancesCloneRequest
Database instance clone request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

clone instances (request)

InstancesExportRequest
Database instance export request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

export instances (request)

InstancesImportRequest
Database instance import request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

import instances (request)

InstancesListResponse
Database instances list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list instances (response)

InstancesRestoreBackupRequest
Database instance restore backup request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

restore backup instances (request)

IpConfiguration
IP Management configuration.

This type is not used in any activity, and only used as part of another schema.

IpMapping
Database instance IP Mapping.

This type is not used in any activity, and only used as part of another schema.

LocationPreference
Preferred location. This specifies where a Cloud SQL instance should preferably be located, either in a specific Compute Engine zone, or co-located with an App Engine application. Note that if the preferred location is not available, the instance will be located as close as possible within the region. Only one location may be specified.

This type is not used in any activity, and only used as part of another schema.

Operation
An Operations resource contains information about database instance operations such as create, delete, and restart. Operations resources are created in response to operations that were initiated; you never create them directly.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete ssl certs (response)

delete users (response)

delete databases (response)

patch databases (response)

reset ssl config instances (response)

promote replica instances (response)

update databases (response)

list operations (none)

update users (response)

insert databases (response)

patch instances (response)

clone instances (response)

delete instances (response)

get operations (response)

stop replica instances (response)

start replica instances (response)

insert users (response)

insert instances (response)

import instances (response)

update instances (response)

restart instances (response)

export instances (response)

restore backup instances (response)

OperationError
Database instance operation error.

This type is not used in any activity, and only used as part of another schema.

OperationErrorNested
If errors occurred during processing of this operation, this field will be populated.

This type is not used in any activity, and only used as part of another schema.

OperationGetMethodBuilder
Retrieves an instance operation that has been performed on an instance.

OperationListMethodBuilder
Lists all instance operations that have been performed on the given Cloud SQL instance in the reverse chronological order of the start time.

OperationMethodsBuilder
A builder providing access to all methods supported on operation resources. It is not used directly, but through the SQLAdmin hub.

OperationsListResponse
Database instance list operations response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list operations (response)

RestoreBackupContext
Database instance restore from backup context.

This type is not used in any activity, and only used as part of another schema.

SQLAdmin
Central instance to access all SQLAdmin related resource activities

Settings
Database instance settings.

This type is not used in any activity, and only used as part of another schema.

SslCert
SslCerts Resource

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert ssl certs (none)

get ssl certs (response)

list ssl certs (none)

delete ssl certs (none)

SslCertDeleteMethodBuilder
Deletes the SSL certificate. The change will not take effect until the instance is restarted.

SslCertDetail
SslCertDetail.

This type is not used in any activity, and only used as part of another schema.

SslCertGetMethodBuilder
Retrieves a particular SSL certificate. Does not include the private key (required for usage). The private key must be saved from the response to initial creation.

SslCertInsertMethodBuilder
Creates an SSL certificate and returns it along with the private key and server certificate authority. The new certificate will not be usable until the instance is restarted.

SslCertListMethodBuilder
Lists all of the current SSL certificates for the instance.

SslCertMethodsBuilder
A builder providing access to all methods supported on sslCert resources. It is not used directly, but through the SQLAdmin hub.

SslCertsInsertRequest
SslCerts insert request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert ssl certs (request)

SslCertsInsertResponse
SslCert insert response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert ssl certs (response)

SslCertsListResponse
SslCerts list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list ssl certs (response)

Tier
A Google Cloud SQL service tier resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tiers (none)

TierListMethodBuilder
Lists all available service tiers for Google Cloud SQL, for example D1, D2. For related information, see Pricing.

TierMethodsBuilder
A builder providing access to all methods supported on tier resources. It is not used directly, but through the SQLAdmin hub.

TiersListResponse
Tiers list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tiers (response)

User
A Cloud SQL user resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete users (none)

update users (request)

list users (none)

insert users (request)

UserDeleteMethodBuilder
Deletes a user from a Cloud SQL instance.

UserInsertMethodBuilder
Creates a new user in a Cloud SQL instance.

UserListMethodBuilder
Lists users in the specified Cloud SQL instance.

UserMethodsBuilder
A builder providing access to all methods supported on user resources. It is not used directly, but through the SQLAdmin hub.

UserUpdateMethodBuilder
Updates an existing user in a Cloud SQL instance.

UsersListResponse
User list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list users (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.