Crate google_sqladmin1_beta4

This documentation was generated from SQL Admin crate version 1.0.0+20160907, where 20160907 is the exact revision of the sqladmin:v1beta4 schema built by the mako code generator v1.0.0.

Everything else about the SQL Admin v1_beta4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

backup runs

delete, get, insert and list

databases

delete, get, insert, list, patch and update

flags

list

instances

clone, delete, export, failover, get, import, insert, list, patch, promote replica, reset ssl config, restart, restore backup, start replica, stop replica and update

operations

get and list

ssl certs

create ephemeral, delete, get, insert and list

tiers

list

users

delete, insert, list and update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.ssl_certs().delete(...).doit()
let r = hub.users().delete(...).doit()
let r = hub.databases().delete(...).doit()
let r = hub.instances().failover(...).doit()
let r = hub.databases().patch(...).doit()
let r = hub.instances().reset_ssl_config(...).doit()
let r = hub.instances().promote_replica(...).doit()
let r = hub.databases().update(...).doit()
let r = hub.operations().list(...).doit()
let r = hub.users().update(...).doit()
let r = hub.databases().insert(...).doit()
let r = hub.backup_runs().delete(...).doit()
let r = hub.instances().patch(...).doit()
let r = hub.instances().clone(...).doit()
let r = hub.instances().delete(...).doit()
let r = hub.operations().get(...).doit()
let r = hub.instances().stop_replica(...).doit()
let r = hub.instances().start_replica(...).doit()
let r = hub.users().insert(...).doit()
let r = hub.instances().insert(...).doit()
let r = hub.instances().import(...).doit()
let r = hub.backup_runs().insert(...).doit()
let r = hub.instances().update(...).doit()
let r = hub.instances().restart(...).doit()
let r = hub.instances().export(...).doit()
let r = hub.instances().restore_backup(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-sqladmin1_beta4 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_sqladmin1_beta4 as sqladmin1_beta4;
use sqladmin1_beta4::User;
use sqladmin1_beta4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = User::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().update(req, "project", "instance", "host", "name")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AclEntry
An entry for an Access Control list.

BackupConfiguration
Database instance backup configuration.

BackupRun
A database instance backup run resource.

BackupRunDeleteCall
Deletes the backup taken by a backup run.

BackupRunGetCall
Retrieves a resource containing information about a backup run.

BackupRunInsertCall
Creates a new backup run on demand. This method is applicable only to Second Generation instances.

BackupRunListCall
Lists all backup runs associated with a given instance and configuration in the reverse chronological order of the enqueued time.

BackupRunMethods
A builder providing access to all methods supported on backupRun resources. It is not used directly, but through the SQLAdmin hub.

BackupRunsListResponse
Backup run list results.

BinLogCoordinates
Binary log coordinates.

CloneContext
Database instance clone context.

Database
A database resource inside a Cloud SQL instance.

DatabaseDeleteCall
Deletes a database from a Cloud SQL instance.

DatabaseFlags
MySQL flags for Cloud SQL instances.

DatabaseGetCall
Retrieves a resource containing information about a database inside a Cloud SQL instance.

DatabaseInsertCall
Inserts a resource containing information about a database inside a Cloud SQL instance.

DatabaseInstance
A Cloud SQL instance resource.

DatabaseInstanceFailoverReplica
The name and status of the failover replica. This property is applicable only to Second Generation instances.

DatabaseListCall
Lists databases in the specified Cloud SQL instance.

DatabaseMethods
A builder providing access to all methods supported on database resources. It is not used directly, but through the SQLAdmin hub.

DatabasePatchCall
Updates a resource containing information about a database inside a Cloud SQL instance. This method supports patch semantics.

DatabaseUpdateCall
Updates a resource containing information about a database inside a Cloud SQL instance.

DatabasesListResponse
Database list response.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ExportContext
Database instance export context.

ExportContextCsvExportOptions
Options for exporting data as CSV.

ExportContextSqlExportOptions
Options for exporting data as SQL statements.

FailoverContext
Database instance failover context.

Flag
A Google Cloud SQL service flag resource.

FlagListCall
List all available database flags for Google Cloud SQL instances.

FlagMethods
A builder providing access to all methods supported on flag resources. It is not used directly, but through the SQLAdmin hub.

FlagsListResponse
Flags list response.

ImportContext
Database instance import context.

ImportContextCsvImportOptions
Options for importing data as CSV.

InstanceCloneCall
Creates a Cloud SQL instance as a clone of the source instance. The API is not ready for Second Generation instances yet.

InstanceDeleteCall
Deletes a Cloud SQL instance.

InstanceExportCall
Exports data from a Cloud SQL instance to a Google Cloud Storage bucket as a MySQL dump file.

InstanceFailoverCall
Failover the instance to its failover replica instance.

InstanceGetCall
Retrieves a resource containing information about a Cloud SQL instance.

InstanceImportCall
Imports data into a Cloud SQL instance from a MySQL dump file in Google Cloud Storage.

InstanceInsertCall
Creates a new Cloud SQL instance.

InstanceListCall
Lists instances under a given project in the alphabetical order of the instance name.

InstanceMethods
A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

InstancePatchCall
Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.. This method supports patch semantics.

InstancePromoteReplicaCall
Promotes the read replica instance to be a stand-alone Cloud SQL instance.

InstanceResetSslConfigCall
Deletes all client certificates and generates a new server SSL certificate for the instance. The changes will not take effect until the instance is restarted. Existing instances without a server certificate will need to call this once to set a server certificate.

InstanceRestartCall
Restarts a Cloud SQL instance.

InstanceRestoreBackupCall
Restores a backup of a Cloud SQL instance.

InstanceStartReplicaCall
Starts the replication in the read replica instance.

InstanceStopReplicaCall
Stops the replication in the read replica instance.

InstanceUpdateCall
Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.

InstancesCloneRequest
Database instance clone request.

InstancesExportRequest
Database instance export request.

InstancesFailoverRequest
Instance failover request.

InstancesImportRequest
Database instance import request.

InstancesListResponse
Database instances list response.

InstancesRestoreBackupRequest
Database instance restore backup request.

IpConfiguration
IP Management configuration.

IpMapping
Database instance IP Mapping.

LocationPreference
Preferred location. This specifies where a Cloud SQL instance should preferably be located, either in a specific Compute Engine zone, or co-located with an App Engine application. Note that if the preferred location is not available, the instance will be located as close as possible within the region. Only one location may be specified.

MaintenanceWindow
Maintenance window. This specifies when a v2 Cloud SQL instance should preferably be restarted for system maintenance puruposes.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

MySqlReplicaConfiguration
Read-replica configuration specific to MySQL databases.

OnPremisesConfiguration
On-premises instance configuration.

Operation
An Operations resource contains information about database instance operations such as create, delete, and restart. Operations resources are created in response to operations that were initiated; you never create them directly.

OperationError
Database instance operation error.

OperationErrors
Database instance operation errors list wrapper.

OperationGetCall
Retrieves an instance operation that has been performed on an instance.

OperationListCall
Lists all instance operations that have been performed on the given Cloud SQL instance in the reverse chronological order of the start time.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the SQLAdmin hub.

OperationsListResponse
Database instance list operations response.

ReplicaConfiguration
Read-replica configuration for connecting to the master.

RestoreBackupContext
Database instance restore from backup context.

SQLAdmin
Central instance to access all SQLAdmin related resource activities

Settings
Database instance settings.

SslCert
SslCerts Resource

SslCertCreateEphemeralCall
Generates a short-lived X509 certificate containing the provided public key and signed by a private key specific to the target instance. Users may use the certificate to authenticate as themselves when connecting to the database.

SslCertDeleteCall
Deletes the SSL certificate. The change will not take effect until the instance is restarted.

SslCertDetail
SslCertDetail.

SslCertGetCall
Retrieves a particular SSL certificate. Does not include the private key (required for usage). The private key must be saved from the response to initial creation.

SslCertInsertCall
Creates an SSL certificate and returns it along with the private key and server certificate authority. The new certificate will not be usable until the instance is restarted.

SslCertListCall
Lists all of the current SSL certificates for the instance.

SslCertMethods
A builder providing access to all methods supported on sslCert resources. It is not used directly, but through the SQLAdmin hub.

SslCertsCreateEphemeralRequest
SslCerts create ephemeral certificate request.

SslCertsInsertRequest
SslCerts insert request.

SslCertsInsertResponse
SslCert insert response.

SslCertsListResponse
SslCerts list response.

Tier
A Google Cloud SQL service tier resource.

TierListCall
Lists all available service tiers for Google Cloud SQL, for example D1, D2. For related information, see Pricing.

TierMethods
A builder providing access to all methods supported on tier resources. It is not used directly, but through the SQLAdmin hub.

TiersListResponse
Tiers list response.

User
A Cloud SQL user resource.

UserDeleteCall
Deletes a user from a Cloud SQL instance.

UserInsertCall
Creates a new user in a Cloud SQL instance.

UserListCall
Lists users in the specified Cloud SQL instance.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the SQLAdmin hub.

UserUpdateCall
Updates an existing user in a Cloud SQL instance.

UsersListResponse
User list response.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.