Struct google_sqladmin1_beta4::DatabaseInstance

pub struct DatabaseInstance {
    pub failover_replica: Option<DatabaseInstanceFailoverReplica>,
    pub replica_names: Option<Vec<String>>,
    pub backend_type: Option<String>,
    pub current_disk_size: Option<String>,
    pub service_account_email_address: Option<String>,
    pub ip_addresses: Option<Vec<IpMapping>>,
    pub database_version: Option<String>,
    pub instance_type: Option<String>,
    pub max_disk_size: Option<String>,
    pub on_premises_configuration: Option<OnPremisesConfiguration>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub ipv6_address: Option<String>,
    pub suspension_reason: Option<Vec<String>>,
    pub region: Option<String>,
    pub server_ca_cert: Option<SslCert>,
    pub master_instance_name: Option<String>,
    pub project: Option<String>,
    pub state: Option<String>,
    pub etag: Option<String>,
    pub replica_configuration: Option<ReplicaConfiguration>,
    pub settings: Option<Settings>,
    pub self_link: Option<String>,
}

A Cloud SQL instance resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch instances (request)

insert instances (request)

get instances (response)

update instances (request)

Fields
failover_replica: Option<DatabaseInstanceFailoverReplica>

The name and status of the failover replica. This property is applicable only to Second Generation instances.

replica_names: Option<Vec<String>>

The replicas of the instance.

backend_type: Option<String>

FIRST_GEN: Basic Cloud SQL instance that runs in a Google-managed container. SECOND_GEN: A newer Cloud SQL backend that runs in a Compute Engine VM. EXTERNAL: A MySQL server that is not managed by Google.

current_disk_size: Option<String>

The current disk usage of the instance in bytes. This property has been deprecated. Users should use the "cloudsql.googleapis.com/database/disk/bytes_used" metric in Cloud Monitoring API instead. Please see https://groups.google.com/d/msg/google-cloud-sql-announce/I_7-F9EBhT0/BtvFtdFeAgAJ for details.

service_account_email_address: Option<String>

The service account email address assigned to the instance. This property is applicable only to Second Generation instances.

ip_addresses: Option<Vec<IpMapping>>

The assigned IP addresses for the instance.

database_version: Option<String>

The database engine type and version. The databaseVersion can not be changed after instance creation. Can be MYSQL_5_5, MYSQL_5_6 or MYSQL_5_7. Defaults to MYSQL_5_6. MYSQL_5_7 is applicable only to Second Generation instances.

instance_type: Option<String>

The instance type. This can be one of the following. CLOUD_SQL_INSTANCE: A Cloud SQL instance that is not replicating from a master. ON_PREMISES_INSTANCE: An instance running on the customer's premises. READ_REPLICA_INSTANCE: A Cloud SQL instance configured as a read-replica.

max_disk_size: Option<String>

The maximum disk size of the instance in bytes.

on_premises_configuration: Option<OnPremisesConfiguration>

Configuration specific to on-premises instances.

kind: Option<String>

This is always sql#instance.

name: Option<String>

Name of the Cloud SQL instance. This does not include the project ID.

ipv6_address: Option<String>

The IPv6 address assigned to the instance. This property is applicable only to First Generation instances.

suspension_reason: Option<Vec<String>>

If the instance state is SUSPENDED, the reason for the suspension.

region: Option<String>

The geographical region. Can be us-central (FIRST_GEN instances only), us-central1 (SECOND_GEN instances only), asia-east1 or europe-west1. Defaults to us-central or us-central1 depending on the instance type (First Generation or Second Generation). The region can not be changed after instance creation.

server_ca_cert: Option<SslCert>

SSL configuration.

master_instance_name: Option<String>

The name of the instance which will act as master in the replication setup.

project: Option<String>

The project ID of the project containing the Cloud SQL instance. The Google apps domain is prefixed if applicable.

state: Option<String>

The current serving state of the Cloud SQL instance. This can be one of the following. RUNNABLE: The instance is running, or is ready to run when accessed. SUSPENDED: The instance is not available, for example due to problems with billing. PENDING_CREATE: The instance is being created. MAINTENANCE: The instance is down for maintenance. FAILED: The instance creation failed. UNKNOWN_STATE: The state of the instance is unknown.

etag: Option<String>

HTTP 1.1 Entity tag for the resource.

replica_configuration: Option<ReplicaConfiguration>

Configuration specific to read-replicas replicating from on-premises masters.

settings: Option<Settings>

The user settings.

self_link: Option<String>

The URI of this resource.