Struct google_sqladmin1_beta4::InstanceMethods

pub struct InstanceMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_sqladmin1_beta4 as sqladmin1_beta4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clone(...)`, `delete(...)`, `export(...)`, `failover(...)`, `get(...)`, `import(...)`, `insert(...)`, `list(...)`, `patch(...)`, `promote_replica(...)`, `reset_ssl_config(...)`, `restart(...)`, `restore_backup(...)`, `start_replica(...)`, `stop_replica(...)` and `update(...)`
// to build up your call.
let rb = hub.instances();

Methods

impl<'a, C, A> InstanceMethods<'a, C, A>

fn failover(&self, request: InstancesFailoverRequest, project: &str, instance: &str) -> InstanceFailoverCall<'a, C, A>

Create a builder to help you perform the following task:

Failover the instance to its failover replica instance.

Arguments

request - No description provided.

project - ID of the project that contains the read replica.

instance - Cloud SQL instance ID. This does not include the project ID.

fn reset_ssl_config(&self, project: &str, instance: &str) -> InstanceResetSslConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all client certificates and generates a new server SSL certificate for the instance. The changes will not take effect until the instance is restarted. Existing instances without a server certificate will need to call this once to set a server certificate.

Arguments

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

fn promote_replica(&self, project: &str, instance: &str) -> InstancePromoteReplicaCall<'a, C, A>

Create a builder to help you perform the following task:

Promotes the read replica instance to be a stand-alone Cloud SQL instance.

Arguments

project - ID of the project that contains the read replica.

instance - Cloud SQL read replica instance name.

fn get(&self, project: &str, instance: &str) -> InstanceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a resource containing information about a Cloud SQL instance.

Arguments

project - Project ID of the project that contains the instance.

instance - Database instance ID. This does not include the project ID.

fn patch(&self, request: DatabaseInstance, project: &str, instance: &str) -> InstancePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.. This method supports patch semantics.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

fn clone(&self, request: InstancesCloneRequest, project: &str, instance: &str) -> InstanceCloneCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a Cloud SQL instance as a clone of the source instance. The API is not ready for Second Generation instances yet.

Arguments

request - No description provided.

project - Project ID of the source as well as the clone Cloud SQL instance.

instance - The ID of the Cloud SQL instance to be cloned (source). This does not include the project ID.

fn delete(&self, project: &str, instance: &str) -> InstanceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a Cloud SQL instance.

Arguments

project - Project ID of the project that contains the instance to be deleted.

instance - Cloud SQL instance ID. This does not include the project ID.

fn stop_replica(&self, project: &str, instance: &str) -> InstanceStopReplicaCall<'a, C, A>

Create a builder to help you perform the following task:

Stops the replication in the read replica instance.

Arguments

project - ID of the project that contains the read replica.

instance - Cloud SQL read replica instance name.

fn start_replica(&self, project: &str, instance: &str) -> InstanceStartReplicaCall<'a, C, A>

Create a builder to help you perform the following task:

Starts the replication in the read replica instance.

Arguments

project - ID of the project that contains the read replica.

instance - Cloud SQL read replica instance name.

fn insert(&self, request: DatabaseInstance, project: &str) -> InstanceInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new Cloud SQL instance.

Arguments

request - No description provided.

project - Project ID of the project to which the newly created Cloud SQL instances should belong.

fn list(&self, project: &str) -> InstanceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists instances under a given project in the alphabetical order of the instance name.

Arguments

project - Project ID of the project for which to list Cloud SQL instances.

fn import(&self, request: InstancesImportRequest, project: &str, instance: &str) -> InstanceImportCall<'a, C, A>

Create a builder to help you perform the following task:

Imports data into a Cloud SQL instance from a MySQL dump file in Google Cloud Storage.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

fn update(&self, request: DatabaseInstance, project: &str, instance: &str) -> InstanceUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

fn restart(&self, project: &str, instance: &str) -> InstanceRestartCall<'a, C, A>

Create a builder to help you perform the following task:

Restarts a Cloud SQL instance.

Arguments

project - Project ID of the project that contains the instance to be restarted.

instance - Cloud SQL instance ID. This does not include the project ID.

fn export(&self, request: InstancesExportRequest, project: &str, instance: &str) -> InstanceExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports data from a Cloud SQL instance to a Google Cloud Storage bucket as a MySQL dump file.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance to be exported.

instance - Cloud SQL instance ID. This does not include the project ID.

fn restore_backup(&self, request: InstancesRestoreBackupRequest, project: &str, instance: &str) -> InstanceRestoreBackupCall<'a, C, A>

Create a builder to help you perform the following task:

Restores a backup of a Cloud SQL instance.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.