Struct google_sqladmin1_beta4::InstanceMethods

pub struct InstanceMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on instance resources. It is not used directly, but through the SQLAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_sqladmin1_beta4 as sqladmin1_beta4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use sqladmin1_beta4::SQLAdmin;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SQLAdmin::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_server_ca(...)`, `clone(...)`, `delete(...)`, `demote_master(...)`, `export(...)`, `failover(...)`, `get(...)`, `import(...)`, `insert(...)`, `list(...)`, `list_server_cas(...)`, `patch(...)`, `promote_replica(...)`, `reset_ssl_config(...)`, `restart(...)`, `restore_backup(...)`, `rotate_server_ca(...)`, `start_replica(...)`, `stop_replica(...)`, `truncate_log(...)` and `update(...)`
// to build up your call.
let rb = hub.instances();

Methods

impl<'a, C, A> InstanceMethods<'a, C, A>

pub fn truncate_log(
&self,
request: InstancesTruncateLogRequest,
project: &str,
instance: &str
) -> InstanceTruncateLogCall<'a, C, A>

Create a builder to help you perform the following task:

Truncate MySQL general and slow query log tables

Arguments

request - No description provided.

project - Project ID of the Cloud SQL project.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn demote_master(
&self,
request: InstancesDemoteMasterRequest,
project: &str,
instance: &str
) -> InstanceDemoteMasterCall<'a, C, A>

Create a builder to help you perform the following task:

Demotes the stand-alone instance to be a Cloud SQL read replica for an external database server.

Arguments

request - No description provided.

project - ID of the project that contains the instance.

instance - Cloud SQL instance name.

pub fn failover(
&self,
request: InstancesFailoverRequest,
project: &str,
instance: &str
) -> InstanceFailoverCall<'a, C, A>

Create a builder to help you perform the following task:

Failover the instance to its failover replica instance.

Arguments

request - No description provided.

project - ID of the project that contains the read replica.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn import(
&self,
request: InstancesImportRequest,
project: &str,
instance: &str
) -> InstanceImportCall<'a, C, A>

Create a builder to help you perform the following task:

Imports data into a Cloud SQL instance from a SQL dump or CSV file in Cloud Storage.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn reset_ssl_config(
&self,
project: &str,
instance: &str
) -> InstanceResetSslConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all client certificates and generates a new server SSL certificate for the instance.

Arguments

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn add_server_ca(
&self,
project: &str,
instance: &str
) -> InstanceAddServerCaCall<'a, C, A>

Create a builder to help you perform the following task:

Add a new trusted Certificate Authority (CA) version for the specified instance. Required to prepare for a certificate rotation. If a CA version was previously added but never used in a certificate rotation, this operation replaces that version. There cannot be more than one CA version waiting to be rotated in.

Arguments

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn promote_replica(
&self,
project: &str,
instance: &str
) -> InstancePromoteReplicaCall<'a, C, A>

Create a builder to help you perform the following task:

Promotes the read replica instance to be a stand-alone Cloud SQL instance.

Arguments

project - ID of the project that contains the read replica.

instance - Cloud SQL read replica instance name.

pub fn get(&self, project: &str, instance: &str) -> InstanceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a resource containing information about a Cloud SQL instance.

Arguments

project - Project ID of the project that contains the instance.

instance - Database instance ID. This does not include the project ID.

pub fn patch(
&self,
request: DatabaseInstance,
project: &str,
instance: &str
) -> InstancePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.. This method supports patch semantics.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn clone(
&self,
request: InstancesCloneRequest,
project: &str,
instance: &str
) -> InstanceCloneCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a Cloud SQL instance as a clone of the source instance.

Arguments

request - No description provided.

project - Project ID of the source as well as the clone Cloud SQL instance.

instance - The ID of the Cloud SQL instance to be cloned (source). This does not include the project ID.

pub fn delete(
&self,
project: &str,
instance: &str
) -> InstanceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a Cloud SQL instance.

Arguments

project - Project ID of the project that contains the instance to be deleted.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn stop_replica(
&self,
project: &str,
instance: &str
) -> InstanceStopReplicaCall<'a, C, A>

Create a builder to help you perform the following task:

Stops the replication in the read replica instance.

Arguments

project - ID of the project that contains the read replica.

instance - Cloud SQL read replica instance name.

pub fn start_replica(
&self,
project: &str,
instance: &str
) -> InstanceStartReplicaCall<'a, C, A>

Create a builder to help you perform the following task:

Starts the replication in the read replica instance.

Arguments

project - ID of the project that contains the read replica.

instance - Cloud SQL read replica instance name.

pub fn insert(
&self,
request: DatabaseInstance,
project: &str
) -> InstanceInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new Cloud SQL instance.

Arguments

request - No description provided.

project - Project ID of the project to which the newly created Cloud SQL instances should belong.

pub fn rotate_server_ca(
&self,
request: InstancesRotateServerCaRequest,
project: &str,
instance: &str
) -> InstanceRotateServerCaCall<'a, C, A>

Create a builder to help you perform the following task:

Rotates the server certificate to one signed by the Certificate Authority (CA) version previously added with the addServerCA method.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn list(&self, project: &str) -> InstanceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists instances under a given project in the alphabetical order of the instance name.

Arguments

project - Project ID of the project for which to list Cloud SQL instances.

pub fn list_server_cas(
&self,
project: &str,
instance: &str
) -> InstanceListServerCaCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all of the trusted Certificate Authorities (CAs) for the specified instance. There can be up to three CAs listed: the CA that was used to sign the certificate that is currently in use, a CA that has been added but not yet used to sign a certificate, and a CA used to sign a certificate that has previously rotated out.

Arguments

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn update(
&self,
request: DatabaseInstance,
project: &str,
instance: &str
) -> InstanceUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates settings of a Cloud SQL instance. Caution: This is not a partial update, so you must include values for all the settings that you want to retain. For partial updates, use patch.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn restart(
&self,
project: &str,
instance: &str
) -> InstanceRestartCall<'a, C, A>

Create a builder to help you perform the following task:

Restarts a Cloud SQL instance.

Arguments

project - Project ID of the project that contains the instance to be restarted.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn export(
&self,
request: InstancesExportRequest,
project: &str,
instance: &str
) -> InstanceExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports data from a Cloud SQL instance to a Cloud Storage bucket as a SQL dump or CSV file.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance to be exported.

instance - Cloud SQL instance ID. This does not include the project ID.

pub fn restore_backup(
&self,
request: InstancesRestoreBackupRequest,
project: &str,
instance: &str
) -> InstanceRestoreBackupCall<'a, C, A>

Create a builder to help you perform the following task:

Restores a backup of a Cloud SQL instance.

Arguments

request - No description provided.

project - Project ID of the project that contains the instance.

instance - Cloud SQL instance ID. This does not include the project ID.