Struct google_sqladmin1_beta4::MySqlReplicaConfiguration

pub struct MySqlReplicaConfiguration {
    pub username: Option<String>,
    pub kind: Option<String>,
    pub password: Option<String>,
    pub ssl_cipher: Option<String>,
    pub ca_certificate: Option<String>,
    pub client_certificate: Option<String>,
    pub master_heartbeat_period: Option<String>,
    pub verify_server_certificate: Option<bool>,
    pub dump_file_path: Option<String>,
    pub connect_retry_interval: Option<i32>,
    pub client_key: Option<String>,
}

Read-replica configuration specific to MySQL databases.

This type is not used in any activity, and only used as part of another schema.

Fields
username: Option<String>

The username for the replication connection.

kind: Option<String>

This is always sql#mysqlReplicaConfiguration.

password: Option<String>

The password for the replication connection.

ssl_cipher: Option<String>

A list of permissible ciphers to use for SSL encryption.

ca_certificate: Option<String>

PEM representation of the trusted CA's x509 certificate.

client_certificate: Option<String>

PEM representation of the slave's x509 certificate.

master_heartbeat_period: Option<String>

Interval in milliseconds between replication heartbeats.

verify_server_certificate: Option<bool>

Whether or not to check the master's Common Name value in the certificate that it sends during the SSL handshake.

dump_file_path: Option<String>

Path to a SQL dump file in Google Cloud Storage from which the slave instance is to be created. The URI is in the form gs://bucketName/fileName. Compressed gzip files (.gz) are also supported. Dumps should have the binlog co-ordinates from which replication should begin. This can be accomplished by setting --master-data to 1 when using mysqldump.

connect_retry_interval: Option<i32>

Seconds to wait between connect retries. MySQL's default is 60 seconds.

client_key: Option<String>

PEM representation of the slave's private key. The corresponsing public key is encoded in the client's certificate.