Struct google_sqladmin1_beta4::User

pub struct User {
    pub project: Option<String>,
    pub instance: Option<String>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub name: Option<String>,
    pub password: Option<String>,
    pub host: Option<String>,
}

A Cloud SQL user resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete users (none)

update users (request)

list users (none)

insert users (request)

Fields

project

The project ID of the project containing the Cloud SQL database. The Google apps domain is prefixed if applicable. Can be omitted for update since it is already specified on the URL.

instance

The name of the Cloud SQL instance. This does not include the project ID. Can be omitted for update since it is already specified on the URL.

kind

This is always sql#user.

etag

HTTP 1.1 Entity tag for the resource.

name

The name of the user in the Cloud SQL instance. Can be omitted for update since it is already specified on the URL.

password

The password for the user.

host

The host name from which the user can connect. For insert operations, host defaults to an empty string. For update operations, host is specified as part of the request URL. The host name cannot be updated after insertion.